requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
requests==2.31.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
spacy==3.7.2